HandGestureCheck sample app
A simple "hand gesture" which always returns the transform of each detected hand, and then visualizes them simply with cubes offset from the hand location. This was a testbed to detect issues with leaving and re-entering an ImmersiveSpace.

With HandGesture 26.0.1, leaving and re-entering the immersive space in this sample app would show updates for one hand for a bit and then switch to the other for a bit.  With 26.0.2, both hands continue to smoothly update.